Moved some static file asset test fixtures into html-bundler_test.ts.
