--- test/html/imports/es6-modules/abc.js
@@ -1,4 +1,0 @@
-import { upcase } from './upcase.js';
-export const A = upcase('a');
-export const B = upcase('b');
-export const C = upcase('c');
