--- test/html/imports/es6-modules/def.js
@@ -1,5 +1,0 @@
-import { X, Y, Z } from './xyz.js';
-const D = X + X;
-const E = Y + Y;
-const F = Z + Z;
-export { D, E, F }
