--- test/html/imports/es6-modules/multiple-external-modules.html
@@ -1,3 +1,0 @@
-<script type="module" src="abc.js"></script>
-<script type="module" src="abc.js"></script>
-<script type="module" src="def.js"></script>
