--- test/html/imports/es6-modules/multiple-inline-modules.html
@@ -1,15 +1,0 @@
-<script type="module">
-  import {A, B} from './abc.js';
-  console.log(A,B);
-</script>
-
-<script type="module">
-  import {B, C} from './abc.js';
-  import {Y} from './xyz.js';
-  console.log(B,C,Y);
-</script>
-
-<script type="module">
-  import {D,F} from './def.js';
-  console.log(D,F);
-</script>
