--- test/html/imports/es6-modules/omgz.js
@@ -1,2 +1,0 @@
-import { upcase } from './upcase.js';
-export const Z = upcase('omgz');
