--- test/html/imports/es6-modules/upcase.js
@@ -1,3 +1,0 @@
-export function upcase(str) {
-  return str.toUpperCase();
-}
