--- test/html/imports/es6-modules/xyz.js
@@ -1,4 +1,0 @@
-import { upcase } from './upcase.js';
-export const X = upcase('x');
-export const Y = upcase('y');
-export const Z = upcase('z');
